refactor(DUP): `handle_heartbeat` in a separate module
Moves `heandleì_heartbeat` away form `impl.ex` into a spearate module.
